patches: linux: x86_64: Add two small "hacks"
The Makefile change resolves the following error when ld is not present:

error: Cannot generate ORC metadata for CONFIG_UNWINDER_ORC=y, please
install libelf-dev, libelf-devel or elfutils-libelf-devel

clang tries to invoke ld because '-fuse-ld' is not present in this
command, which we provide via $(HOSTLDFLAGS). I don't know of a better
way to do this right now.

The tools/objtool/Makefile change is because AR via the command line is
not respected. This change can be upstreamed properly I think, I just
need to add a proper commit message :^)

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>