driver: Support booting via EFI on arm64 and x86_64
If the ovmf or qemu-efi-aarch64 packages are installed on Debian or
Ubuntu, use those files to boot up via EFI. This will help prevent
regressions like https://github.com/ClangBuiltLinux/linux/issues/634.

I considered checking in the fd files so that other distributions could
take advantage of this but these files are over 50MB, which is too much
of a burden to force on everyone.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>